+++ client/src/components/Button/Button.tsx
@@ -1,49 +1,8 @@
 import React from 'react';
-import styled, { css } from 'styled-components';
+import { CustomButton } from './Button.styles';
+import { ButtonProps } from './Button.types';
 
-interface IButtonProps {
-  text: string;
-  clickHandler: (
-    evt:
-      | React.MouseEvent<HTMLButtonElement>
-      | React.KeyboardEvent<HTMLButtonElement>
-  ) => void;
-  disabled: boolean;
-  ariaLabel: string;
-}
-
-const CustomButton = styled.button<{ disabled: boolean }>`
-  width: 100%;
-  height: 50px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 5px;
-  background: ${({ disabled }) => (disabled ? 'darkgray' : 'white')};
-  color: black;
-  font-size: 24px;
-  font-family: inherit;
-  letter-spacing: 10px;
-  cursor: ${({ disabled }) => (disabled ? 'not-allowed' : 'pointer')};
-  transform: scale(1);
-  transition: all 0.5s;
-  ${({ disabled }) => {
-    return (
-      !disabled &&
-      css`
-        :hover {
-          transform: scale(1.05);
-        }
-
-        :click {
-          transform: scale(0.95);
-        }
-      `
-    );
-  }}
-`;
-
-export const Button: React.FC<IButtonProps> = ({
+export const Button: React.FC<ButtonProps> = ({
   text,
   clickHandler,
   disabled,
