+++ client/src/components/Button/Button.types.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export interface ButtonProps {
+  text: string;
+  clickHandler: (
+    evt:
+      | React.MouseEvent<HTMLButtonElement>
+      | React.KeyboardEvent<HTMLButtonElement>
+  ) => void;
+  disabled: boolean;
+  ariaLabel: string;
+}
