+++ client/src/components/Input/Input.types.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+export interface InputProps {
+  name: string;
+  labelId: string;
+  inputId: string;
+  type: CustomInputType;
+  label: string;
+  blurFn: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  required?: boolean;
+  changeFn?: (errored: boolean) => void;
+  validator?: CustomInputType;
+  defaultColor?: string;
+}
+
+export type CustomInputType =
+  | 'email'
+  | 'money'
+  | 'number'
+  | 'password'
+  | 'phone'
+  | 'text'
+  | 'zip';
