+++ client/src/components/Loader/Loader.types.tsx
@@ -1,0 +1,3 @@
+export interface LoaderProps {
+  loading: boolean;
+}
