+++ client/src/components/Login/Login.styles.tsx
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+
+export const LoginWrapper = styled.main`
+  background: black;
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  .sign-up {
+    color: white;
+    display: flex;
+
+    span {
+      color: lightblue;
+      cursor: pointer;
+    }
+  }
+`;
