+++ client/src/components/Login/Login.types.tsx
@@ -1,0 +1,3 @@
+export interface LoginProps {
+  setUserAuthed: (bool: boolean) => void;
+}
