+++ client/src/components/MainDash/MainDash.types.tsx
@@ -1,0 +1,3 @@
+export interface MainDashProps {
+  setUserAuthed: (bool: boolean) => void;
+}
