+++ client/src/components/Modal/Modal.types.tsx
@@ -1,0 +1,5 @@
+export interface ModalProps {
+  children: any;
+  toggle: () => void;
+  open: boolean;
+}
