+++ client/src/components/Portal/Portal.tsx
@@ -1,0 +1,7 @@
+import { createPortal } from 'react-dom';
+
+export const Portal = ({ children }: any) => {
+  const modalRoot = document.getElementById('modal');
+  if (modalRoot) return createPortal(children, modalRoot);
+  return null;
+};
