+++ client/src/components/SignUp/SignUp.styles.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const SignUpWrapper = styled.div`
+  .errored {
+    border-bottom: solid 5px red;
+  }
+`;
