+++ client/src/components/SignUp/SignUp.types.tsx
@@ -1,0 +1,5 @@
+export interface SignUpProps {
+  handleSignUpSuccess: () => void;
+  triggerLoader: (trigger: boolean) => void;
+  duplicateEmailTriggered: () => void;
+}
